requirements.txt
# SPDX-FileCopyrightText: 2025 Mel0nABC
#
# SPDX-License-Identifier: MIT
astroid==3.3.11
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
dill==0.4.0
et_xmlfile==2.0.0
fonttools==4.58.5
gbulb==0.6.6
isort==6.0.1
kiwisolver==1.4.8
lxml==6.0.0
markdown-it-py==3.0.0
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.17.0
mypy_extensions==1.1.0
numpy==2.3.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
path==17.1.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pycairo==1.28.0
Pygments==2.19.2
PyGObject==3.48.2
pylint==3.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
rich==14.1.0
shellingham==1.5.4
six==1.17.0
tomlkit==0.13.3
typer==0.12.5
typing_extensions==4.14.1
watchdog==6.0.0
